Move non-component fhir evaluation code to its own service
