+++ containers/ecr-viewer/src/app/tests/utils.test.tsx
@@ -1,10 +1,6 @@
 import {
   DisplayDataProps,
-  evaluateSocialData,
-  extractPatientAddress,
-  calculatePatientAge,
   evaluateClinicalData,
-  evaluatePatientName,
   returnProblemsTable,
   returnCareTeamTable,
   isDataAvailable,
@@ -27,6 +23,12 @@ import { CarePlanActivity } from "fhir/r4b";
 import { evaluate } from "fhirpath";
 import userEvent from "@testing-library/user-event";
 import { Tooltip } from "@trussworks/react-uswds";
+import {
+  evaluateSocialData,
+  evaluatePatientName,
+  evaluatePatientAddress,
+  calculatePatientAge,
+} from "../services/evaluateFhirDataService";
 
 describe("Utils", () => {
   const mappings = loadYamlConfig();
@@ -109,12 +111,12 @@ describe("Utils", () => {
   });
   describe("Extract Patient Address", () => {
     it("should return empty string if no address is available", () => {
-      const actual = extractPatientAddress(undefined as any, mappings);
+      const actual = evaluatePatientAddress(undefined as any, mappings);
 
       expect(actual).toBeEmpty();
     });
     it("should get patient address", () => {
-      const actual = extractPatientAddress(
+      const actual = evaluatePatientAddress(
         BundleWithPatient as unknown as Bundle,
         mappings,
       );
