+++ containers/ecr-viewer/src/app/utils.tsx
@@ -1,5 +1,4 @@
 import React, { ReactNode, useEffect, useState } from "react";
-import * as dateFns from "date-fns";
 import {
   Bundle,
   Condition,
@@ -14,21 +13,22 @@ import parse from "html-react-parser";
 import classNames from "classnames";
 
 import {
-  formatAddress,
   formatDate,
   formatName,
-  formatPhoneNumber,
-  formatStartEndDateTime,
   formatVitals,
   formatTablesToJSON,
   TableRow,
   removeHtmlElements,
   toSentenceCase,
 } from "@/app/services/formatService";
-import { evaluateTable, evaluateReference } from "./services/evaluateService";
+import { evaluateTable } from "./services/evaluateService";
 import { Button, Table } from "@trussworks/react-uswds";
 import { CareTeamParticipant, CarePlanActivity } from "fhir/r4b";
 import { Tooltip } from "@trussworks/react-uswds";
+import {
+  calculatePatientAge,
+  evaluateReference,
+} from "./services/evaluateFhirDataService";
 
 export interface DisplayDataProps {
   title?: string;
@@ -59,361 +59,6 @@ export interface CompleteData {
 export const noData = (
   <span className="no-data text-italic text-base">No data</span>
 );
-
-/**
- * Evaluates patient name from the FHIR bundle and formats it into structured data for display.
- * @param fhirBundle - The FHIR bundle containing patient contact info.
- * @param mappings - The object containing the fhir paths.
- * @returns The formatted patient name
- */
-export const evaluatePatientName = (
-  fhirBundle: Bundle,
-  mappings: PathMappings,
-) => {
-  const givenNames = evaluate(fhirBundle, mappings.patientGivenName).join(" ");
-  const familyName = evaluate(fhirBundle, mappings.patientFamilyName);
-
-  return `${givenNames} ${familyName}`;
-};
-
-/**
- * Evaluates patient address from the FHIR bundle and formats it into structured data for display.
- * @param fhirBundle - The FHIR bundle containing patient contact info.
- * @param mappings - The object containing the fhir paths.
- * @returns The formatted patient address
- */
-export const extractPatientAddress = (
-  fhirBundle: Bundle,
-  mappings: PathMappings,
-) => {
-  const streetAddresses = evaluate(fhirBundle, mappings.patientStreetAddress);
-  const city = evaluate(fhirBundle, mappings.patientCity)[0];
-  const state = evaluate(fhirBundle, mappings.patientState)[0];
-  const zipCode = evaluate(fhirBundle, mappings.patientZipCode)[0];
-  const country = evaluate(fhirBundle, mappings.patientCountry)[0];
-  return formatAddress(streetAddresses, city, state, zipCode, country);
-};
-
-/**
- * Extracts a specific location resource from a given FHIR bundle based on defined path mappings.
- * @param fhirBundle - The FHIR bundle object containing various resources, including location resources.
- * @param fhirPathMappings - An object containing FHIR path mappings, which should include a mapping
- *   for `facilityLocation` that determines how to find the location reference within the bundle.
- * @returns The location resource object from the FHIR bundle that matches the UID derived from the
- *   facility location reference. If no matching resource is found, the function returns `undefined`.
- */
-function extractLocationResource(
-  fhirBundle: Bundle,
-  fhirPathMappings: PathMappings,
-) {
-  const locationReference = evaluate(
-    fhirBundle,
-    fhirPathMappings.facilityLocation,
-  ).join("");
-  const locationUID = locationReference.split("/")[1];
-  const locationExpression = `Bundle.entry.resource.where(resourceType = 'Location').where(id = '${locationUID}')`;
-  return evaluate(fhirBundle, locationExpression)[0];
-}
-
-/**
- * Evaluates facility address from the FHIR bundle and formats it into structured data for display.
- * @param fhirBundle - The FHIR bundle containing patient contact info.
- * @param mappings - The object containing the fhir paths.
- * @returns The formatted facility address
- */
-export const extractFacilityAddress = (
-  fhirBundle: Bundle,
-  mappings: PathMappings,
-) => {
-  const locationResource = extractLocationResource(fhirBundle, mappings);
-
-  const streetAddresses = locationResource?.address?.line;
-  const city = locationResource?.address?.city;
-  const state = locationResource?.address?.state;
-  const zipCode = locationResource?.address?.postalCode;
-  const country = locationResource?.address?.country;
-
-  return formatAddress(streetAddresses, city, state, zipCode, country);
-};
-
-/**
- * Evaluates patient contact info from the FHIR bundle and formats it into structured data for display.
- * @param fhirBundle - The FHIR bundle containing patient contact info.
- * @param mappings - The object containing the fhir paths.
- * @returns All phone numbers and emails seperated by new lines
- */
-export const evaluatePatientContactInfo = (
-  fhirBundle: Bundle,
-  mappings: PathMappings,
-) => {
-  const phoneNumbers = evaluate(fhirBundle, mappings.patientPhoneNumbers)
-    .map(
-      (phoneNumber) =>
-        `${
-          phoneNumber?.use?.charAt(0).toUpperCase() +
-          phoneNumber?.use?.substring(1)
-        } ${phoneNumber.value}`,
-    )
-    .join("\n");
-  const emails = evaluate(fhirBundle, mappings.patientEmails)
-    .map((email) => `${email.value}`)
-    .join("\n");
-
-  return `${phoneNumbers}\n${emails}`;
-};
-
-/**
- * Extracts travel history information from the provided FHIR bundle based on the FHIR path mappings.
- * @param fhirBundle - The FHIR bundle containing patient travel history data.
- * @param mappings - An object containing the FHIR path mappings.
- * @returns - A formatted string representing the patient's travel history, or undefined if no relevant data is found.
- */
-const extractTravelHistory = (
-  fhirBundle: Bundle,
-  mappings: PathMappings,
-): string | undefined => {
-  const startDate = evaluate(
-    fhirBundle,
-    mappings["patientTravelHistoryStartDate"],
-  )[0];
-  const endDate = evaluate(
-    fhirBundle,
-    mappings["patientTravelHistoryEndDate"],
-  )[0];
-  const location = evaluate(
-    fhirBundle,
-    mappings["patientTravelHistoryLocation"],
-  )[0];
-  const purposeOfTravel = evaluate(
-    fhirBundle,
-    mappings["patientTravelHistoryPurpose"],
-  )[0];
-  if (startDate || endDate || location || purposeOfTravel) {
-    return `Dates: ${startDate} - ${endDate}
-       Location(s): ${location ?? "No data"}
-       Purpose of Travel: ${purposeOfTravel ?? "No data"}
-       `;
-  }
-};
-
-/**
- * Calculates the age of a patient to a given date or today.
- * @param fhirBundle - The FHIR bundle containing patient information.
- * @param fhirPathMappings - The mappings for retrieving patient date of birth.
- * @param [givenDate] - Optional. The target date to calculate the age. Defaults to the current date if not provided.
- * @returns - The age of the patient in years, or undefined if date of birth is not available.
- */
-export const calculatePatientAge = (
-  fhirBundle: Bundle,
-  fhirPathMappings: PathMappings,
-  givenDate?: string,
-) => {
-  const patientDOBString = evaluate(fhirBundle, fhirPathMappings.patientDOB)[0];
-  if (patientDOBString) {
-    const patientDOB = new Date(patientDOBString);
-    const targetDate = givenDate ? new Date(givenDate) : new Date();
-    return dateFns.differenceInYears(targetDate, patientDOB);
-  }
-};
-
-/**
- * Evaluates social data from the FHIR bundle and formats it into structured data for display.
- * @param fhirBundle - The FHIR bundle containing social data.
- * @param mappings - The object containing the fhir paths.
- * @returns An array of evaluated and formatted social data.
- */
-export const evaluateSocialData = (
-  fhirBundle: Bundle,
-  mappings: PathMappings,
-) => {
-  const socialData: DisplayDataProps[] = [
-    {
-      title: "Occupation",
-      value: evaluate(fhirBundle, mappings["patientCurrentJobTitle"])[0],
-    },
-    {
-      title: "Tobacco Use",
-      value: evaluate(fhirBundle, mappings["patientTobaccoUse"])[0],
-    },
-    {
-      title: "Travel History",
-      value: extractTravelHistory(fhirBundle, mappings),
-    },
-    {
-      title: "Homeless Status",
-      value: evaluate(fhirBundle, mappings["patientHomelessStatus"])[0],
-    },
-    {
-      title: "Pregnancy Status",
-      value: evaluate(fhirBundle, mappings["patientPregnancyStatus"])[0],
-    },
-    {
-      title: "Alcohol Use",
-      value: evaluate(fhirBundle, mappings["patientAlcoholUse"])[0],
-    },
-    {
-      title: "Sexual Orientation",
-      value: evaluate(fhirBundle, mappings["patientSexualOrientation"])[0],
-    },
-    {
-      title: "Gender Identity",
-      value: evaluate(fhirBundle, mappings["patientGenderIdentity"])[0],
-    },
-    {
-      title: "Occupation",
-      value: evaluate(fhirBundle, mappings["patientCurrentJobTitle"])[0],
-    },
-  ];
-  return evaluateData(socialData);
-};
-
-/**
- * Evaluates demographic data from the FHIR bundle and formats it into structured data for display.
- * @param fhirBundle - The FHIR bundle containing demographic data.
- * @param mappings - The object containing the fhir paths.
- * @returns An array of evaluated and formatted demographic data.
- */
-export const evaluateDemographicsData = (
-  fhirBundle: Bundle,
-  mappings: PathMappings,
-) => {
-  const demographicsData: DisplayDataProps[] = [
-    {
-      title: "Patient Name",
-      value: evaluatePatientName(fhirBundle, mappings),
-    },
-    { title: "DOB", value: evaluate(fhirBundle, mappings.patientDOB)[0] },
-    {
-      title: "Current Age",
-      value: calculatePatientAge(fhirBundle, mappings)?.toString(),
-    },
-    {
-      title: "Vital Status",
-      value: evaluate(fhirBundle, mappings.patientVitalStatus)[0]
-        ? "Deceased"
-        : "Alive",
-    },
-    { title: "Sex", value: evaluate(fhirBundle, mappings.patientGender)[0] },
-    { title: "Race", value: evaluate(fhirBundle, mappings.patientRace)[0] },
-    {
-      title: "Ethnicity",
-      value: evaluate(fhirBundle, mappings.patientEthnicity)[0],
-    },
-    {
-      title: "Tribal Affiliation",
-      value: evaluate(fhirBundle, mappings.patientTribalAffiliation)[0],
-    },
-    {
-      title: "Preferred Language",
-      value: evaluate(fhirBundle, mappings.patientLanguage)[0],
-    },
-    {
-      title: "Patient Address",
-      value: extractPatientAddress(fhirBundle, mappings),
-    },
-    {
-      title: "County",
-      value: evaluate(fhirBundle, mappings.patientCounty)[0],
-    },
-    {
-      title: "Contact",
-      value: evaluatePatientContactInfo(fhirBundle, mappings),
-    },
-    {
-      title: "Emergency Contact",
-      value: evaluateEmergencyContact(fhirBundle, mappings),
-    },
-    {
-      title: "Patient IDs",
-      toolTip:
-        "Unique patient identifier(s) from their medical record. For example, a patient's social security number or medical record number.",
-      value: evaluate(fhirBundle, mappings.patientId)[0],
-    },
-  ];
-  return evaluateData(demographicsData);
-};
-
-/**
- * Evaluates encounter data from the FHIR bundle and formats it into structured data for display.
- * @param fhirBundle - The FHIR bundle containing encounter data.
- * @param mappings - The object containing the fhir paths.
- * @returns An array of evaluated and formatted encounter data.
- */
-export const evaluateEncounterData = (
-  fhirBundle: Bundle,
-  mappings: PathMappings,
-) => {
-  const encounterData = [
-    {
-      title: "Encounter Date/Time",
-      value: formatStartEndDateTime(
-        evaluate(fhirBundle, mappings["encounterStartDate"])[0],
-        evaluate(fhirBundle, mappings["encounterEndDate"])[0],
-      ),
-    },
-    {
-      title: "Encounter Type",
-      value: evaluate(fhirBundle, mappings["encounterType"])[0],
-    },
-    {
-      title: "Facility Name",
-      value: evaluate(fhirBundle, mappings["facilityName"])[0],
-    },
-    {
-      title: "Facility Address",
-      value: formatAddress(
-        evaluate(fhirBundle, mappings["facilityStreetAddress"]),
-        evaluate(fhirBundle, mappings["facilityCity"])[0],
-        evaluate(fhirBundle, mappings["facilityState"])[0],
-        evaluate(fhirBundle, mappings["facilityZipCode"])[0],
-        evaluate(fhirBundle, mappings["facilityCountry"])[0],
-      ),
-    },
-    {
-      title: "Facility Contact",
-      value: formatPhoneNumber(
-        evaluate(fhirBundle, mappings["facilityContact"])[0],
-      ),
-    },
-    {
-      title: "Facility Type",
-      value: evaluate(fhirBundle, mappings["facilityType"])[0],
-    },
-    {
-      title: "Facility ID",
-      value: evaluate(fhirBundle, mappings["facilityID"])[0],
-    },
-  ];
-  return evaluateData(encounterData);
-};
-
-/**
- * Evaluates provider data from the FHIR bundle and formats it into structured data for display.
- * @param fhirBundle - The FHIR bundle containing provider data.
- * @param mappings - The object containing the fhir paths.
- * @returns An array of evaluated and formatted provider data.
- */
-export const evaluateProviderData = (
-  fhirBundle: Bundle,
-  mappings: PathMappings,
-) => {
-  const providerData = [
-    {
-      title: "Provider Name",
-      value: formatName(
-        evaluate(fhirBundle, mappings["providerGivenName"]),
-        evaluate(fhirBundle, mappings["providerFamilyName"])[0],
-      ),
-    },
-    {
-      title: "Provider Contact",
-      value: formatPhoneNumber(
-        evaluate(fhirBundle, mappings["providerContact"])[0],
-      ),
-    },
-  ];
-  return evaluateData(providerData);
-};
 
 /**
  * Generates a formatted table representing the list of problems based on the provided array of problems and mappings.
@@ -1191,56 +836,6 @@ const trimField = (
     return { value: newElement, remainingLength: split.remainingLength };
   }
   return { value, remainingLength: remainingLength };
-};
-
-/**
- * Evaluates emergency contact information from the FHIR bundle and formats it into a readable string.
- * @param fhirBundle - The FHIR bundle containing patient information.
- * @param mappings - The object containing the fhir paths.
- * @returns The formatted emergency contact information.
- */
-export const evaluateEmergencyContact = (
-  fhirBundle: Bundle,
-  mappings: PathMappings,
-) => {
-  const contact = evaluate(fhirBundle, mappings.patientEmergencyContact)[0];
-
-  let formattedContact;
-
-  if (contact) {
-    if (contact.relationship) {
-      const relationship = contact.relationship;
-      formattedContact = `${relationship}`;
-    }
-
-    if (contact.address) {
-      const address = formatAddress(
-        contact.address[0].line,
-        contact.address[0].city,
-        contact.address[0].state,
-        contact.address[0].postalCode,
-        contact.address[0].country,
-      );
-
-      formattedContact = `${formattedContact}\n${address}`;
-    }
-
-    if (contact.telecom) {
-      const phoneNumbers = evaluate(fhirBundle, mappings.patientPhoneNumbers)
-        .map(
-          (phoneNumber) =>
-            `${
-              phoneNumber?.use?.charAt(0).toUpperCase() +
-              phoneNumber?.use?.substring(1)
-            } ${phoneNumber.value}`,
-        )
-        .join("\n");
-
-      formattedContact = `${formattedContact}\n${phoneNumbers}`;
-    }
-
-    return formattedContact;
-  }
 };
 
 type CustomDivProps = React.PropsWithChildren<{
