+++ containers/ecr-viewer/src/app/view-data/components/AccordionContainer.tsx
@@ -1,11 +1,4 @@
-import {
-  evaluateSocialData,
-  evaluateEncounterData,
-  evaluateProviderData,
-  evaluateClinicalData,
-  evaluateDemographicsData,
-  PathMappings,
-} from "../../utils";
+import { evaluateClinicalData, PathMappings } from "../../utils";
 import Demographics from "./Demographics";
 import SocialHistory from "./SocialHistory";
 import UnavailableInfo from "./UnavailableInfo";
@@ -19,6 +12,12 @@ import LabInfo from "@/app/view-data/components/LabInfo";
 import { formatString } from "@/app/services/formatService";
 import { evaluateEcrMetadata } from "../../services/ecrMetadataService";
 import { evaluateLabInfoData } from "@/app/services/labsService";
+import {
+  evaluateDemographicsData,
+  evaluateSocialData,
+  evaluateEncounterData,
+  evaluateProviderData,
+} from "@/app/services/evaluateFhirDataService";
 
 type AccordionContainerProps = {
   children?: ReactNode;
